Simplify backend to a single improved server entry
