+++ package.json
@@ -5,9 +5,7 @@
   "main": "backend/server.js",
   "scripts": {
     "start": "node backend/server.js",
-    "start:improved": "set USE_IMPROVED_API=true&& node backend/server-improved.js",
     "dev": "nodemon backend/server.js",
-    "dev:improved": "set USE_IMPROVED_API=true&& nodemon backend/server-improved.js",
     "test": "jest --coverage",
     "test:watch": "jest --watch",
     "lint": "eslint .",
@@ -52,7 +50,6 @@
     "uuid": "^9.0.1"
   },
   "devDependencies": {
-    "cross-env": "^10.1.0",
     "eslint": "^8.54.0",
     "jest": "^29.7.0",
     "nodemon": "^3.0.1",
